[feat][#29][#28] Add option to choose internal classifier and rename to classification-with-embeddings
