+++ README.md
@@ -1,2 +1,2 @@
-# text-classification-with-embeddings
-Evaluation of entity embedding methods on tasks of text classification
+# classification-with-embeddings
+Evaluation of entity embedding methods on tasks of classification
