+++ __main__.py
@@ -1,13 +1,17 @@
 import argparse
 
-from text_classification_with_embeddings import Tasks, EntityEmbeddingMethod
-from text_classification_with_embeddings import logger
-from text_classification_with_embeddings.document_embedding import embed
-from text_classification_with_embeddings.document_embedding.embed import get_word_to_embedding
-from text_classification_with_embeddings.evaluation.evaluate import evaluate
-from text_classification_with_embeddings.evaluation.get_clf import get_clf_with_internal_clf, get_clf_starspace
-from text_classification_with_embeddings.evaluation.train_test_split import get_train_test_split
-from text_classification_with_embeddings.util.argparse import file_path, dir_path, proportion_float
+from sklearn.ensemble import RandomForestClassifier
+from sklearn.linear_model import LogisticRegression
+from sklearn.svm import SVC
+
+from classification_with_embeddings import Tasks, EntityEmbeddingMethod, InternalClassifier
+from classification_with_embeddings import logger
+from classification_with_embeddings.embedding import embed
+from classification_with_embeddings.embedding.embed import get_word_to_embedding
+from classification_with_embeddings.evaluation.evaluate import evaluate
+from classification_with_embeddings.evaluation.get_clf import get_clf_with_internal_clf, get_clf_starspace
+from classification_with_embeddings.evaluation.train_test_split import get_train_test_split
+from classification_with_embeddings.util.argparse import file_path, dir_path, proportion_float
 
 
 def main(**kwargs):
@@ -30,17 +34,17 @@ def main(**kwargs):
 def task_get_entity_embeddings(**kwargs):
     if kwargs['method'] == EntityEmbeddingMethod.STARSPACE.value:
         # STARSPACE
-        logger.info('using StarSpace method')
+        logger.info('Using StarSpace method')
         if kwargs['starspace_path'] is None:
             raise ValueError('path to StarSpace executable must be defined when using StarSpace')
         embed.get_starspace_entity_embeddings(kwargs['starspace_path'], kwargs['train_data_path'], kwargs['output_dir_path'], kwargs['starspace_args'])
     elif kwargs['method'] == EntityEmbeddingMethod.WORD2VEC.value:
         # WORD2VEC
-        logger.info('using Word2Vec method')
+        logger.info('Using Word2Vec method')
         embed.get_word2vec_embeddings(kwargs['train_data_path'], kwargs['output_dir_path'], kwargs['word2vec_args'])
     elif kwargs['method'] == EntityEmbeddingMethod.FASTTEXT.value:
         # FASTTEXT
-        logger.info('using fastText method')
+        logger.info('Using fastText method')
         embed.get_fasttext_embeddings(kwargs['train_data_path'], kwargs['output_dir_path'], kwargs['fasttext_args'])
     else:
         raise NotImplementedError('Method {0} not implemented'.format(kwargs['method']))
@@ -59,7 +63,17 @@ def task_evaluate(**kwargs):
     elif kwargs['method'] == EntityEmbeddingMethod.WORD2VEC.value or kwargs['method'] == EntityEmbeddingMethod.FASTTEXT.value:
         # WORD2VEC or FASTTEXT
         logger.info('Evaluating {0} method'.format('Word2Vec' if kwargs['method'] == EntityEmbeddingMethod.WORD2VEC.value else 'fastText'))
-        clf = get_clf_with_internal_clf(word_to_embedding, kwargs['train_data_path'], None, kwargs['internal_clf_args'])
+
+        # set internal classifier
+        if kwargs['internal_clf'] == InternalClassifier.LOGISTIC_REGRESSION.value:
+            clf_internal = LogisticRegression
+        elif kwargs['internal_clf'] == InternalClassifier.RANDOM_FOREST.value:
+            clf_internal = RandomForestClassifier
+        elif kwargs['internal_clf'] == InternalClassifier.SVC.value:
+            clf_internal = SVC
+        else:
+            raise NotImplementedError('Classifier {0} not implemented'.format(kwargs['internal_clf']))
+        clf = get_clf_with_internal_clf(word_to_embedding, kwargs['train_data_path'], clf_internal, kwargs['internal_clf_args'])
     else:
         raise NotImplementedError('Method {0} not implemented'.format(kwargs['method']))
     evaluate(clf, kwargs['method'], kwargs['test_data_path'], kwargs['results_path'])
@@ -67,7 +81,7 @@ def task_evaluate(**kwargs):
 
 if __name__ == '__main__':
     # parse arguments
-    parser = argparse.ArgumentParser(prog='text-classification-with-embeddings')
+    parser = argparse.ArgumentParser(prog='classification-with-embeddings')
     subparsers = parser.add_subparsers(required=True, dest='task', help='Task to run')
 
     # COMPUTING DOCUMENT EMBEDDINGS FROM fastText FORMAT INPUT
@@ -99,6 +113,7 @@ if __name__ == '__main__':
     evaluate_parser.add_argument('--test-data-path', type=file_path, required=True, help='Path to file containing the test data in fastText format')
     evaluate_parser.add_argument('--embeddings-path', type=file_path, required=True, help='Path to stored feature embeddings')
     evaluate_parser.add_argument('--results-path', type=dir_path, default='.', help='Path to directory in which to save the results')
+    evaluate_parser.add_argument('--internal-clf', type=str, default=InternalClassifier.LOGISTIC_REGRESSION.value, help='Internal classifier to use (if applicable)')
     evaluate_parser.add_argument('--internal-clf-args', type=str, default='',
                                  help='Arguments passed to internal classifier if applicable (key-value pairs such as val=1 enclose in quotes with no commas separated by spaces)')
